pyproject.toml
[project]
name = "assignment1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.3",
    "fastapi>=0.128.3",
    "pydantic-settings>=2.12.0",
    "scalar-fastapi>=1.6.1",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]
